+++ src/sonarr_metadata_rewrite/file_scanner.py
@@ -6,7 +6,7 @@ from collections.abc import Callable
 from pathlib import Path
 
 from sonarr_metadata_rewrite.config import Settings
-from sonarr_metadata_rewrite.nfo_utils import find_target_files
+from sonarr_metadata_rewrite.file_utils import find_target_files
 
 logger = logging.getLogger(__name__)
